requirements.txt
apache-beam[gcp]==2.41.0
numpy==1.23.5
pandas==1.5.2
google-cloud-bigquery==3.9.0
google-api-python-client==2.86.0
google-cloud-storage==2.8.0
pytest-xdist==3.1.0
pytest==7.0.1
pyyaml==6.0